Prevent errors when using fallback to `fetch` mechanism.
Prior to this change, if none of the registered fetch handlers return
something other than `undefined` we would hit this gaurd to resolve with
the result of fetching the asset (which is good). Unfortunately, in that
case, we would also continue execution and attempt to call
`handler.then` when `handler` is known to be undefined.

This change prevents that by returning early once we have resolved the
fallback scenario.