Don't run sw.js through babel in non-minification envs
Babel will cache sw.js and that cache cannot be busted unless the
original file on the file system changes. Because sw.js is not on the
file system it will never change. This means in development mode the
cache busting version string can never be updated.

This PR will move babel's transpilation step into the uglify step. Babel
is only required to satisfy uglify requiring ES5 for the time being.